+++ src/plone/restapi/tests/test_documentation.py
@@ -6,6 +6,7 @@ from datetime import timedelta
 from freezegun import freeze_time
 from mock import patch
 from pkg_resources import parse_version
+from pkg_resources import resource_filename
 from plone import api
 from plone.app.discussion.interfaces import IConversation
 from plone.app.discussion.interfaces import IDiscussionSettings
@@ -71,14 +72,8 @@ RESPONSE_HEADER_KEYS = [
     'location',
 ] + TUS_HEADERS
 
-base_path = os.path.join(
-    os.path.dirname(__file__),
-    '..',
-    '..',
-    '..',
-    '..',
-    'docs/source/_json'
-)
+
+base_path = resource_filename('plone.restapi', '../../../docs/source/_json')
 
 UPLOAD_DATA = 'abcdefgh'
 UPLOAD_MIMETYPE = 'text/plain'
@@ -219,6 +214,19 @@ class TestDocumentation(unittest.TestCase):
     def tearDown(self):
         self.time_freezer.stop()
         popGlobalRegistry(getSite())
+
+    def test_http_example_path_exists(self):
+        """Tripwire test to ensure changes in the HTTP examples base path
+
+        are caught *before* release time, since setup.py relies on it.
+        """
+        path = resource_filename('plone.restapi', '../../../docs/source/_json')
+        self.assertTrue(
+            os.path.isdir(path),
+            'It looks like the directory for the HTTP examples has changed. '
+            'Please make sure to update HTTP_EXAMPLES_PATH in setup.py '
+            'accordingly, as well as base_path at the top of this file and '
+            'the one in this test.')
 
     def test_documentation_content_crud(self):
         folder = self.create_folder()